Moved user input handling into custom MinesweeperMouseAdapter
